requirements.txt
astroid==2.3.3
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
prompt-toolkit==3.0.5
pycryptodome==3.9.7
pycryptodomex==3.9.7
pylint==2.4.4
requests==2.23.0
six==1.14.0
urllib3==1.25.9
wcwidth==0.1.9
wrapt==1.11.2
